Move llm.mapper tests to tests/bofire/llm/
Mirrors the kernels/ and priors/ conventions where mapper tests live next
to the module they exercise, not in the strategies suite.
